backend/requirements.txt
# Core
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic-settings==2.0.3
pydantic[email]==2.4.2

# Database
supabase==2.0.2
psycopg2-binary==2.9.9

# Data processing
pandas==2.1.3
numpy==1.26.2
openpyxl==3.1.2

# Scraping
aiohttp==3.9.1
playwright==1.40.0
beautifulsoup4==4.12.2
selenium==4.16.0
undetected-chromedriver==3.5.4

# AI & ML
openai==0.28.1

# Utils
python-dotenv==1.0.0
pytest==7.4.3
pytest-asyncio==0.21.1